cpu: Separate difftest flow and initialize before traps
