+++ src/cpu/SConscript
@@ -122,6 +122,7 @@ Source('thread_context.cc')
 Source('thread_state.cc')
 Source('timing_expr.cc')
 Source('difftest.cc')
+Source('difftest_cpu.cc')
 
 SimObject('DummyChecker.py', sim_objects=['DummyChecker'])
 Source('checker/cpu.cc')
